+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTRMH8E8S6STZV6BPEFF585W"
+        "01KTS4FJFCHQ7H6PDVMTY2AKMP"
     )
     r"""The unique identifier of the data source"""