+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTRMH8EEBTYY7JBD63ESSEKJ"
+        "01KTS4FJFHFEPSR4WX708J32HH"
     )
     r"""The unique identifier of the data source"""