+++ packages/orq-rc/RELEASES.md
@@ -8609,3 +8609,13 @@ Based on:
 - [python v4.11.0-rc.10] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.10] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.10 - packages/orq-rc
+
+## 2026-06-10 16:07:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.775.0 (2.900.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.11] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.11] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.11 - packages/orq-rc